authCtrl.js 일부 수정, route/my폴더 생성 및 mypage.js 생성 외
