+++ ctrl/authCtrl.js
@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken')
 const crypto = require('crypto')
 const nodemailer = require('nodemailer')
 
-const { User, Order, Follow } = require('../models')
+const { User } = require('../models')
 const { sendMail } = require('../routes/utils/mailer') // 컨픽에서 메일 전송 함수 호출
 
 const SECRET = process.env.JWT_SECRET || 'minimart-secret-key'
@@ -84,140 +84,6 @@ exports.logout = async (req, res) => {
    }
 }
 
-exports.getMe = async (req, res) => {
-   try {
-      // 1. 유저 정보
-      const user = await User.findByPk(req.user.id, {
-         attributes: ['id', 'email', 'name', 'address', 'phone_number', 'profile_img', 'provider', 'role', 'createdAt'],
-      })
-
-      if (!user) {
-         return res.status(404).json({ message: '사용자를 찾을 수 없습니다.' })
-      }
-
-      // 2. 주문 내역 (Order + Product 조인)
-      const orders = await Order.findAll({
-         where: { user_id: req.user.id },
-         include: [
-            {
-               model: Product,
-               attributes: ['name', 'image_url'],
-            },
-         ],
-         order: [['createdAt', 'DESC']],
-      })
-
-      const formattedOrders = orders.map((order) => ({
-         order_id: order.id,
-         product_name: order.Product.name,
-         product_image: order.Product.image_url,
-         order_date: order.createdAt.toISOString().split('T')[0],
-         status: order.status,
-      }))
-
-      // 3. 팔로잉한 판매자 목록 (Follow + User 조인)
-      const follows = await Follow.findAll({
-         where: { follower_id: req.user.id },
-         include: [
-            {
-               model: User,
-               as: 'Following',
-               attributes: ['id', 'name', 'profile_img'],
-            },
-         ],
-      })
-
-      const formattedFollowings = follows.map((f) => ({
-         seller_id: f.Following.id,
-         seller_name: f.Following.name,
-         seller_profile_img: f.Following.profile_img,
-      }))
-
-      // 4. 응답
-      res.status(200).json({
-         success: true,
-         data: {
-            user: {
-               ...user.toJSON(),
-               createdAt: user.createdAt.toISOString().split('T')[0],
-            },
-            orders: formattedOrders,
-            followings: formattedFollowings,
-         },
-      })
-   } catch (error) {
-      console.error(error)
-      res.status(500).json({ message: '서버 에러' })
-   }
-}
-
-// 내 정보 수정
-exports.updateMe = async (req, res) => {
-   try {
-      const userId = req.user.id // middlewares에서 토큰 검증을 거쳐 붙은 값
-      const { name, address, phone_number, profile_img } = req.body
-
-      const user = await User.findByPk(userId)
-      if (!user) {
-         return res.status(404).json({ message: '사용자를 찾을 수 없습니다.' })
-      }
-
-      user.name = name || user.name
-      user.address = address || user.address
-      user.phone_number = phone_number || user.phone_number
-      user.profile_img = profile_img || user.profile_img
-
-      await user.save()
-
-      res.status(200).json({
-         message: '회원 정보가 수정되었습니다.',
-         user: {
-            id: user.id,
-            email: user.email,
-            address: user.address,
-            phone_number: user.phone_number,
-            profile_img: user.profile_img,
-         },
-      })
-   } catch (error) {
-      console.error(error)
-      res.status(500).json({ message: '서버 에러' })
-   }
-}
-
-// 회원 탈퇴
-exports.deleteAccount = async (req, res) => {
-   try {
-      const userId = req.user.id
-
-      // 1. 사용자 DB에서 삭제
-      await User.destroy({ where: { id: userId } })
-
-      // 2. 세션이 있을 경우 세션 삭제
-      if (req.session) {
-         req.session.destroy((err) => {
-            if (err) {
-               console.error('세션 삭제 실패:', err)
-               return res.status(500).json({ message: '회원 탈퇴는 되었지만, 세션 삭제에 실패했습니다.' })
-            }
-         })
-      }
-
-      // 3. 클라이언트 쿠키 삭제
-      res.clearCookie('connect.sid')
-
-      res.status(200).json({ message: '회원 탈퇴 성공' })
-   } catch (error) {
-      console.error('회원 탈퇴 오류:', error)
-      res.status(500).json({ message: '서버 오류로 인해 회원 탈퇴에 실패했습니다.' })
-   }
-}
-
-// 자동 로그인
-exports.autoLogin = async (req, res) => {
-   res.status(200).json({ message: '자동 로그인 성공', user: req.user })
-}
-
 // 이메일로 비밀번호 초기화 - 인증코드 전송
 exports.sendEmailCode = async (req, res) => {
    const { email } = req.body
@@ -292,11 +158,6 @@ exports.resetPwByEmail = async (req, res) => {
       console.error('비밀번호 변경 실패:', error)
       return res.status(500).json({ message: '서버 에러' })
    }
-}
-
-// 구글 소셜 로그인
-exports.googleLogin = (req, res) => {
-   res.send('구글 로그인')
 }
 
 // 판매자 자격 신청
